+++ hutool-core/src/main/java/cn/hutool/core/map/MapUtil.java
@@ -1546,27 +1546,6 @@ public class MapUtil {
 	}
 
 	/**
-	 * 递归调用将多层级Map处理为一个层级Map类型
-	 *
-	 * @param map 入参Map
-	 * @param flatMap 单层级Map返回值
-	 * @param <K>  键类型
-	 * @param <V>  值类型
-	 */
-	private static <K, V>  void flatten(Map<K, V> map, Map<K, V> flatMap) {
-		for (Map.Entry<K, V> entry : map.entrySet()) {
-			K key = entry.getKey();
-			V value = entry.getValue();
-			if (value instanceof Map) {
-				flatten((Map<K, V>) value, flatMap);
-			} else {
-				flatMap.put(key, value);
-			}
-		}
-	}
-
-
-	/**
 	 * 将多层级Map处理为一个层级Map类型
 	 *
 	 * @param map 入参Map
@@ -1574,10 +1553,35 @@ public class MapUtil {
 	 * @param <K>  键类型
 	 * @param <V>  值类型
 	 */
-	public static <K, V> Map<K, V> flatten(Map<K, V> map) {
+	public static <K, V> Map<K, V> flatten(final Map<K, V> map) {
+		return flatten(map, new HashMap<>());
+	}
+
+	/**
+	 * 递归调用将多层级Map处理为一个层级Map类型
+	 *
+	 * @param map     入参Map
+	 * @param flatMap 单层级Map返回值
+	 * @param <K>     键类型
+	 * @param <V>     值类型
+	 * @return 单层级Map返回值
+	 */
+	@SuppressWarnings("unchecked")
+	public static <K, V> Map<K, V> flatten(final Map<K, V> map, Map<K, V> flatMap) {
 		Assert.notNull(map);
-		Map<K, V> flatMap = new HashMap<>();
-		flatten(map, flatMap);
+		if(null == flatMap){
+			flatMap = new HashMap<>();
+		}
+
+		Map<K, V> finalFlatMap = flatMap;
+		map.forEach((k, v) -> {
+			if (v instanceof Map) {
+				flatten((Map<K, V>) v, finalFlatMap);
+			} else {
+				finalFlatMap.put(k, v);
+			}
+		});
+
 		return flatMap;
 	}
 }