+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 * 【captcha】      `MathGenerator`四则运算方式支持不生成负数结果（pr#1363@Gitee）
 * 【core   】      增加`MapValueProvider`和`RecordConverter`并支持Record转换（issue#3985@Github）
 * 【core   】      `CalendarUtil`增加`isSameYear`和`calendar`方法（issue#3995@Github）
-* 【core   】      `DateUtil`新增方法`yyyy-MM-dd'T'HH:mmXXX`格式支持（pr#1367@Gitee）
+* 【core   】      `DateUtil`增加`yyyy-MM-dd'T'HH:mmXXX`格式支持（pr#1367@Gitee）
+* 【core   】      `MapUtil`增加flatten方法（pr#1368@Gitee）
 
 ### 🐞Bug修复
 * 【extra  】      `Sftp``reconnectIfTimeout`方法改为捕获所有异常（issue#3989@Github）