+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * 【extra  】      修复`QLExpressEngine`allowClassSet无效问题（issue#3994@Github）
 * 【core   】      修复`StrBuilder`insert插入计算错误问题（issue#ICTSRZ@Gitee）
 * 【cron   】      修复`CronPatternUtil.nextDateAfter`计算下一个匹配表达式的日期时，计算错误问题（issue#4006@Github）
+* 【cach   】      修复`AbstrachCache.get`可能造成的死锁问题（issue#4022@Github）
 
 -------------------------------------------------------------------------------------------------------------
 # 5.8.39(2025-06-20)