+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * 【core   】      `MapUtil`增加`removeByValue`和`removeIf`方法
 * 【core   】      `ObjectUtil`增加`apply`方法
 * 【core   】      `ReferenceUtil`增加`get`方法
+* 【db     】      `Condition`增加构造方法支持BETWEEN（issue#4041@Github）
 
 ### 🐞Bug修复
 * 【core   】      修复`ReflectUtil`中因class和Method关联导致的缓存无法回收问题（issue#4039@Github）